Feat: Seperate Data From GameController & Rename Files
monobehavior 상속받지 않은 C# 클래스로 분리해보려 했으나 역시 어렵다.,..
