+++ Assets/Scripts/Runtime/CH1/Pacmom/PMData.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: bf07c5c2bf023084aa688c27fcc41120
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
